Move creation tests to MLS module
